backend/pyproject.toml
[project]
name = "interview-agent-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[anthropic]",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
]
